add a couple of explicit casts to int

+++ preprocessing/ct_batch.py
@@ -233,7 +233,7 @@ class CTImagesBatch(Batch): # pylint: disable=too-many-public-methods
         """
         # set images-component to 3d-array of zeroes if the component is to be updated
         if 'images' in kwargs['src']: 
-            shapes = np.zeros((len(self), 3))
+            shapes = np.zeros((len(self), 3), dtype=np.int)
             for ix in self.indices:
                 filename = os.path.join(self.index.get_fullpath(ix), 'shape.pkl')
                 ix_pos = self._get_verified_pos(ix)
@@ -248,7 +248,7 @@ class CTImagesBatch(Batch): # pylint: disable=too-many-public-methods
             self.images = np.zeros(skysc_shape)
 
             # update bounds of items
-            self._bounds = np.cumsum(np.insert(shapes[:, 0], 0, 0))
+            self._bounds = np.cumsum(np.insert(shapes[:, 0], 0, 0), dtype=np.int)
 
         return self.indices
